+++ src/coreclr/src/pal/tests/palsuite/threading/WaitForMultipleObjectsEx/test2/test2.cpp
@@ -146,7 +146,7 @@ DWORD PALAPI WaiterProc(LPVOID lpParameter)
             GetLastError());
     }
 
-    Alertable = (BOOL) lpParameter;
+    Alertable = (BOOL)(SIZE_T) lpParameter;
 
     LARGE_INTEGER performanceFrequency;
     if (!QueryPerformanceFrequency(&performanceFrequency))
